+++ homeassistant/components/fritz/coordinator.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable, Mapping, ValuesView
+from collections.abc import Callable, Mapping
 from dataclasses import dataclass, field
 from datetime import datetime, timedelta
 from functools import partial
@@ -47,6 +47,7 @@ from .const import (
     FRITZ_EXCEPTIONS,
     MeshRoles,
 )
+from .helpers import _ha_is_stopping
 from .models import (
     ConnectionInfo,
     Device,
@@ -70,35 +71,6 @@ class FritzData:
     tracked: dict[str, set[str]] = field(default_factory=dict)
     profile_switches: dict[str, set[str]] = field(default_factory=dict)
     wol_buttons: dict[str, set[str]] = field(default_factory=dict)
-
-
-def _is_tracked(mac: str, current_devices: ValuesView[set[str]]) -> bool:
-    """Check if device is already tracked."""
-    return any(mac in tracked for tracked in current_devices)
-
-
-def device_filter_out_from_trackers(
-    mac: str,
-    device: FritzDevice,
-    current_devices: ValuesView[set[str]],
-) -> bool:
-    """Check if device should be filtered out from trackers."""
-    reason: str | None = None
-    if device.ip_address == "":
-        reason = "Missing IP"
-    elif _is_tracked(mac, current_devices):
-        reason = "Already tracked"
-
-    if reason:
-        _LOGGER.debug(
-            "Skip adding device %s [%s], reason: %s", device.hostname, mac, reason
-        )
-    return bool(reason)
-
-
-def _ha_is_stopping(activity: str) -> None:
-    """Inform that HA is stopping."""
-    _LOGGER.warning("Cannot execute %s: HomeAssistant is shutting down", activity)
 
 
 class ClassSetupMissing(Exception):
